Reorganize stories
