+++ javascript/react/src/components/ArqueroDetailsList/__tests__/ArqueroDetailsList.stories.tsx
@@ -3,41 +3,17 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 /* eslint-disable @typescript-eslint/consistent-type-imports */
-import { PerformanceTestStory } from './PerformanceTestStory/PerformanceTestStory.js'
-import { RowGroupingTestStory } from './RowGroupingTestStory/RowGroupingTestStory.js'
+import { ParameterizedStory } from './Parameterized.story.js'
+import { PerformanceStory } from './Performance/Performance.story.js'
+import { RowGroupingStory } from './RowGrouping/RowGrouping.story.js'
 
-const meta = {
+export default {
 	title: 'ArqueroDetailsList',
 }
 
-export default meta
+export const RowGrouping = RowGroupingStory
 
-export const RowGroupingStory = (
-	_args: unknown,
-	{ loaded: { stocks } }: any,
-): JSX.Element => {
-	if (!stocks) {
-		return <div>Loading</div>
-	}
+export const Performance = PerformanceStory
+Performance.storyName = 'Performance Test (slow!)'
 
-	return <RowGroupingTestStory mockTable={stocks} />
-}
-
-RowGroupingStory.story = {
-	name: 'Row grouping',
-}
-
-export const PerformanceStory = (
-	_args: unknown,
-	{ loaded: { stocks } }: any,
-): JSX.Element => {
-	if (!stocks) {
-		return <div>Loading...</div>
-	}
-
-	return <PerformanceTestStory mockTablePerformance={stocks} />
-}
-
-PerformanceStory.story = {
-	name: 'Performance',
-}
+export const Parameterized = ParameterizedStory
