+++ javascript/react/src/components/ArqueroDetailsList/__tests__/RowGrouping/RowGrouping.styles.ts
