--- javascript/react/src/components/ArqueroDetailsList/__tests__/SharedStyles.styles.ts
@@ -1,11 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-
-import styled from '@essex/styled-components'
-
-export const Table = styled.div`
-	margin-top: 12px;
-	height: calc(100vh - 220px);
-`
